refactor(migration): add record migrations
